requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.7.10
python-dotenv==1.0.1